Change namespace of tracker sample Move ObjectTrackRenderer to separate file
