+++ samples/ObjectTracker/cs/ObjectTrackerSample_UWP/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application
-    x:Class="ObjectTrackerSample.App"
+    x:Class="ObjectTrackerSkillSample.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:ObjectTrackerSample">
+    xmlns:local="using:ObjectTrackerSkillSample">
 
 </Application>
