+++ samples/ObjectTracker/cs/ObjectTrackerSample_UWP/App.xaml.cs
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-namespace ObjectTrackerSample
+namespace ObjectTrackerSkillSample
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
