+++ samples/ObjectTracker/cs/ObjectTrackerSample_UWP/MainPage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="ObjectTrackerSample.MainPage"
+    x:Class="ObjectTrackerSkillSample.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:ObjectTrackerSample"
+    xmlns:local="using:ObjectTrackerSkillSample"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
