+++ samples/ObjectTracker/cs/ObjectTrackerSample_UWP/ObjectTrackerSample_UWP.csproj
@@ -138,6 +138,7 @@
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
+    <Compile Include="ObjectTrackRenderer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
